requirements.txt
Django==1.11
django-jet==1.0.7
djangorestframework==3.8.2
